Refactor status command into separate module
Move status command implementation from main.rs to commands/status.rs
for consistency with other commands (fix, review, path, clean).

Changes:
- Create src/commands/status.rs with all status logic
- Move MinionStatus struct and helper functions
- Move status tests to status module
- Update main.rs to route to status::handle_status()

This reduces main.rs from ~240 to 94 lines and improves code organization.
