+++ src/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod clean;
 pub mod fix;
 pub mod path;
 pub mod review;
+pub mod status;
